src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Di Huynh <di.huynh179@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9"
pygithub = { version = "^1.58.2", python = ">=3.7" }
urllib3 = "<2"


[tool.poetry.group.test.dependencies]
opentelemetry-distro = {extras = ["otlp"], version = "^0.39b0"}
opentelemetry-sdk-extension-aws = "^2.0.1"
opentelemetry-propagator-aws-xray = "^1.0.1"
boto3 = "^1.26.139"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
